Separate OpenACC kernels regions in data and parallel parts
This is the first in a series of patches that completely rework the handling
of the OpenACC "kernels" directive. In the future, kernels regions will be
transformed into data regions containing a sequence of serial and parallel
offloaded regions. This first patch sets up a new pass that is responsible
for this transformation, and in a first step constructs the new data region
containing a parallel region with the original kernels region's body.

    gcc/
    * Makefile.in: Add...
    * omp-oacc-kernels.c: ... this new file for the kernels conversion pass.
    * passes.def: Add pass_convert_oacc_kernels to pipeline.
    * tree-pass.h (make_pass_convert_oacc_kernels): Add declaration.

    gcc/testsuite/
    * c-c++-common/goacc/kernels-conversion.c: New test.
    * gfortran.dg/goacc/kernels-conversion.f95: Likewise.