New OpenACC kernels region decompose algorithm
Previously, OpenACC kernels region bodies were decomposed into a sequence of
alternating gang-single and gang-parallel "parallel" regions. The new
algorithm in this patch introduces a third possibility: Loops that look like
they might benefit from the parloops pass are converted into old "kernels"
regions, exposing them to the parloops pass later on. This has the benefit
that loops that cannot be parallelized are not offloaded to the GPU.

    gcc/
    * omp-oacc-kernels.c (make_gang_single_region): Check "region sequence
    code", add debug diagnostic code.
    (is_unconditional_oacc_for_loop): Remove stmt parameter and check.
    (determine_region_sequence_code_walk_stmt_fn): New function.
    (determine_region_sequence_code): New function with two overloads.
    (decompose_kernels_region_body): Update for possibility of generating
    kernels regions.

    gcc/testsuite/
    * c-c++-common/goacc/kernels-conversion.c: Adjust test.
    * gfortran.dg/goacc/kernels-conversion.f95: Likewise.